@@ -1,13 +1,15 @@
 import React from 'react'
 import { Provider } from 'redux-bundler-react'
-import { resultSummary as SorghumSummary } from 'sorghum-search'
+import SorghumSummary from '../../sorghum-search/resultSummary'
 // import { resultSummary as GrameneSummary } from 'gramene-search'
 
-export default (store) => (
+const Summary = (store) => (
   <Provider store={store}>
     <ul className="list-unstyled category">
       <SorghumSummary />
       {/*<GrameneSummary />*/}
     </ul>
   </Provider>
-)
+);
+
+export default Summary;
